extracted readypicker from roundrobinloadbalancer
